refactor: extract channel message handlers from engine/mod.rs into engine/channel_handler.rs
Moves the 8 handler functions (handle_channel_message, handle_status_command,
handle_stats_command, handle_subscribe_command, handle_unsubscribe_command,
handle_stream_command, send_channel_reply, forward_to_tmux) into a new
src/engine/channel_handler.rs module.

engine/mod.rs shrinks from ~1915 to ~1324 lines; channel_handler.rs is 609 lines.
No behaviour changes — pure structural move.
